Moving connection.allocate_ids function to datastore.__init__.
