+++ gcloud/datastore/test___init__.py
@@ -205,8 +205,34 @@ class Test_implicit_behavior(unittest2.TestCase):
             result = gcloud.datastore.get_entities(DUMMY_KEYS)
         self.assertTrue(result == DUMMY_VALS)
 
+
+class Test_allocate_ids_function(unittest2.TestCase):
+
+    def _callFUT(self, incomplete_key, num_ids,
+                 connection=None, dataset_id=None):
+        from gcloud.datastore import allocate_ids
+        return allocate_ids(incomplete_key, num_ids, connection=connection,
+                            dataset_id=dataset_id)
+
     def test_allocate_ids(self):
-        import gcloud.datastore
+        from gcloud.datastore.key import Key
+        from gcloud.datastore.test_dataset import _Connection
+
+        DATASET_ID = 'DATASET'
+        INCOMPLETE_KEY = Key('KIND', dataset_id=DATASET_ID)
+        CONNECTION = _Connection()
+        NUM_IDS = 2
+        result = self._callFUT(INCOMPLETE_KEY, NUM_IDS,
+                               connection=CONNECTION, dataset_id=DATASET_ID)
+
+        # Check the IDs returned match.
+        self.assertEqual([key.id for key in result], range(NUM_IDS))
+
+        # Check connection is called correctly.
+        self.assertEqual(CONNECTION._called_dataset_id, DATASET_ID)
+        self.assertEqual(len(CONNECTION._called_key_pbs), NUM_IDS)
+
+    def test_allocate_ids_implicit(self):
         from gcloud.datastore import _implicit_environ
         from gcloud.datastore.key import Key
         from gcloud.datastore.test_dataset import _Connection
@@ -219,7 +245,22 @@ class Test_implicit_behavior(unittest2.TestCase):
         with _Monkey(_implicit_environ, DATASET=CUSTOM_DATASET,
                      CONNECTION=CUSTOM_CONNECTION):
             INCOMPLETE_KEY = Key('KIND')
-            result = gcloud.datastore.allocate_ids(INCOMPLETE_KEY, NUM_IDS)
+            result = self._callFUT(INCOMPLETE_KEY, NUM_IDS)
 
         # Check the IDs returned.
         self.assertEqual([key.id for key in result], range(NUM_IDS))
+
+    def test_allocate_ids_with_complete(self):
+        from gcloud.datastore import _implicit_environ
+        from gcloud.datastore.key import Key
+        from gcloud.datastore.test_dataset import _Connection
+        from gcloud.datastore.test_entity import _Dataset
+        from gcloud._testing import _Monkey
+
+        CUSTOM_DATASET = _Dataset()
+        CUSTOM_CONNECTION = _Connection()
+        with _Monkey(_implicit_environ, DATASET=CUSTOM_DATASET,
+                     CONNECTION=CUSTOM_CONNECTION):
+            COMPLETE_KEY = Key('KIND', 1234)
+            self.assertRaises(ValueError, self._callFUT,
+                              COMPLETE_KEY, 2)
